[restructure] move TableExporter to lib/utils module
